extract abstract class for gen. model #32
