+++ src/main/java/cryptator/gen/CryptaGenBaseModel.java
@@ -9,99 +9,26 @@
 package cryptator.gen;
 
 import org.chocosolver.solver.Model;
-import org.chocosolver.solver.variables.BoolVar;
-import org.chocosolver.solver.variables.IntVar;
-
-import cryptator.specs.ICryptaGenModel;
 
 /**
  * Base class for generation model.
  */
-public class CryptaGenBaseModel implements ICryptaGenModel {
+public abstract class CryptaGenBaseModel extends AbstractCryptaGenModel {
 
-    /** The model. */
-    protected final Model model;
-
-    /** The word array . */
-    protected final String[] words;
-
-    /** The word variables that indicates if the word is present. */
-    protected final BoolVar[] vwords;
-
-    /** The word count. */
-    protected final IntVar wordCount;
-
-    /** The maximum length of a present word. */
-    protected final IntVar maxLength;
-
-    public CryptaGenBaseModel(final Model model, final String[] words, final String prefix,
+    protected CryptaGenBaseModel(final Model model, final String[] words, final String prefix,
             final boolean boundedDomain) {
-        super();
-        this.model = model;
-        this.words = words;
-        this.vwords = buildWordVars(model, words, prefix);
-        this.wordCount = model.intVar(prefix + "wordCount", 0, words.length);
-        this.maxLength = model.intVar(prefix + "maxLength", 0, GenerateUtil.getMaxLength(words), boundedDomain);
+        super(model, words, prefix, boundedDomain);
+
     }
 
     @Override
-    public final Model getModel() {
-        return model;
+    protected void postWordConstraints() {
+        // Do nothing
     }
 
     @Override
-    public final String[] getWords() {
-        return words;
-    }
-
-    @Override
-    public final BoolVar[] getWordVars() {
-        return vwords;
-    }
-
-    @Override
-    public final IntVar getWordCount() {
-        return wordCount;
-    }
-
-    @Override
-    public final IntVar getMaxLength() {
-        return maxLength;
-    }
-
-    /**
-     * Builds named boolean variables associated to the words.
-     *
-     * @param model  the model
-     * @param words  the words list
-     * @param prefix the prefix for the variable names
-     * @return the array of boolean variables
-     */
-    private static BoolVar[] buildWordVars(final Model model, final String[] words, final String prefix) {
-        final BoolVar[] vars = new BoolVar[words.length];
-        for (int i = 0; i < words.length; i++) {
-            vars[i] = model.boolVar(prefix + words[i]);
-        }
-        return vars;
-    }
-
-    protected void postWordCountConstraint() {
-        model.sum(vwords, "=", wordCount).post();
-    }
-
     protected void postMaxLengthConstraints() {
         maxLength.eq(0).decompose().post();
-    }
-
-    @Override
-    public void buildModel() {
-        postWordCountConstraint();
-        postMaxLengthConstraints();
-    }
-
-    @Override
-    public String toString() {
-        return GenerateUtil.recordString(this, " ");
     }
 
 }
